requirements.txt
cudatoolkit==10.0.130
h5py==3.2.1
opencv==4.5.3
pandas==1.3.2
pytorch==1.4.0
scikit-image==0.18.3
scipy==1.7.1
tensorboard==2.4.0
torchaudio==0.4.0
torchvision==0.5.0
tqdm==4.62.2
